Добавляю возможности к функционалу будильников. Пока тестируется

+++ domain/src/main/kotlin/dev/fabled/domain/use_cases/alarms/CreateNewAlarm.kt
@@ -23,7 +23,7 @@ class CreateNewAlarm @Inject constructor(
 
         when (result) {
             is Resource.Success -> {
-                applicationUtil.setupAlarm(alarmModel, result.data)
+                applicationUtil.setupAlarm(alarmModel, result.data.toInt())
             }
 
             else -> emit(result)
